requirements.txt
# Core ML
torch>=2.0
transformers>=4.40
datasets>=2.19
sentencepiece
protobuf
accelerate

# CTranslate2 (optional NLLB optimization)
ctranslate2

# API clients
cohere>=5.0
openai>=1.30
anthropic>=0.25
google-generativeai>=0.5
boto3
azure-ai-translation-text
deepl

# Quality metrics
unbabel-comet>=2.2
sacrebleu>=2.4
bert-score>=0.3.13
fast-langdetect>=0.2

# Infrastructure
tenacity>=8.2
aiolimiter>=1.1
pydantic>=2.7
pyyaml>=6.0
python-dotenv>=1.0
structlog>=24.1
rich>=13.7
huggingface-hub>=0.23

# Testing
pytest>=8.0
pytest-asyncio>=0.23
